cost function graphs

+++ app/src/model/types.h
@@ -1,1 +1,1 @@
-typedef double rk_real;
+typedef float rk_real;
